Create a new Message class
This Message class has a print method, which can be called
to print the message in a specific format. In this case, the
format is to wrap the message in horizontal lines.

Let's implement the Message class into YihuiBot so that
the messages print by YihuiBot uses the new Message class.

Remove wrapStringWithHorizontalLines method in YihuiBot
as that method is no longer needed.