Updated Storage.java
The list inside storage only stores Strings which are primitive structures.

Primitive Structures does not allow for complex representation of data which is what we eventually want to achieve with the Todos, Events and Deadlines in this project.

Thus, the list structure in the storage class needs to be modified and the accompanying methods also need to be fixed accordingly to allow a more complex data structure,  the Task class, to be stored in the list.

This allows each item inside the list to be packaged with both the task name and the mark status together in one object, and eventually we can extend this to other information we want to store together instead of separately.