Renamed Storage.java to TaskList.java
The name storage should be reserved for dealing with loading and saving tasks in the file.

The current structure inside the file TaskList.java contains the task list itself and has operations to add/delete tasks in the list.

I will eventually separate the two but since this file already contains the tasklist itself, i will proceed to keep rename it as TaskList.java.

This will prevent further confusion in future and is more intuitive for the reader of this code to follow.