Added AI.md
I forsee the use of AI to autocomplete quite alot of methods which are similar to each other.

Many of the methods which I forsee creating are kind of repetitive and time consuming in the next few levels.

As such I have created AI.md to explain my rationale for using AI tools in a widespread manner and the explanations behind each use.

This will allow me to document my AI tools used concurrently with the progress of the project, to ensure adherence to the rules of the use of AI in the project.