Added Parser.java
There is many different parts of the code dealing with making sense of the user input.

This makes the debugging process and the code itself very messy.

In order to modulate the code better, I have created Parser.java to only handle all of the user input and change it to a state that the program can understand.

This will allow centralised handling of user inputs and better management of commands by the program.