Refactor code
Create a new Executable class.

The current implementation of YihuiBot results in too much logic written
in YihuiBot.java.

Let's implement the Executable class. Instance of this class can be
executed by YihuiBot. Calling Executable.execute() will execute it
based on what type of executable it is. Example of Executables are AddTask
and ListTask.

Create a new Parser class.

The parser will parse user's input, and create the appropriate executable
so that it can be executed by YihuiBot.