Added FormattedPrinting.java
There is too many printing instances in many different files.

The formatting of the output is not always consistent and the handling of the printing is very messy for the coder.

Lets create a new printing class to handle all of the outputs from the chatbot. Ensure that the class is static so
that new instances will not have to be called and the printing can be done by invoking the class name only.

Using a static class to handle all the printing ensures that there is only one operation of printing
and multiple instances of printing does not occur within the program, leading to messy outputs.
It also makes it easier to use the printer anywhere without having to instantiate it leading to
cleaner code and debugging.