Add Deadline.java, Events.java, ToDo.java
The Task class has the common behaviour for all the different types
of items that can be found inside of the TaskList inside the chatbot.

However, it cannot capture all possible types of behaviours of the
many different types of tasks such as ToDos, Deadlines and Events
which all need to be configured differently.

Let's use polymorphism to extend the task class to these 3 different
classes. Let's change Task class to an abstract class with an abstract
method "getDesctiption" which is to be implemented differently by
each different type of Task. The constructor will also be shortened
to allow each type of Task to store the description differently.
Lastly, lets change the CommandExecutor file to call these different
types of tasks under each different types of commands.

Using polymorphism allows for the 3 different classes to have common
behaviours that do not need to be recoded at any minor change while
also be called using a common object type (Task) especially when we
are not sure what type of object will be returned. This also gives
us flexibility to change the behaviour of each class independently
from one another.