Add CommandTypes.java
Long switch statement inside of Parser.java is not ideal.

It needs to be manually modified every time a new comment is added.
This means we are repeating the change every time we were to add a
new command both to the Enum List inside the CommandExecutor file
and the Parser file.

Lets create a separate Enum file to abstract out all the possible
commands and create a method inside Parser.java to search for the
command inside this Enum class.

Using enum's is preferable than a bunch of switch statements as it
reduces the amount of recoding necessary when new commands are
implemented and it also makes the code look cleaner to readers.