Add Error Handling
The program is not built to handle errors correctly.

This causes the program to crash if there is an input error by the
user, thus the program is not able to provide feedback to the user
about their input.

Let's add the following error handling functionality to the chatbot:
- marking/unmarking error: when the user tries to mark something
  that has already been marked/unmarked
- trying to view or mark items in an empty list
- adding a task without anything
- adding incorrect number of inputs for todo, deadline and events
  respectively.
- entering non-integers after mark/unmark
- marking/unmarking items out of the bounds of the list.

With the following error handling, the chatbot becomes more robust
and is able to provide useful feedback to the user about their input
so that the user is able to self correct their mistakes.