Add delete task functionality
There is no way to delete tasks after you are done with them.

This makes the list look very messy after a few tasks have been
completed.

Let's create a delete functionality to allow users to delete tasks.

This will make the list easier to organise for the user, especially
if they have completed a task or added the task wrongly.