Parser class: Add isDelete() method and delete logic in handleCommand method
This is to check if the command contains a delete, which will initiate the logic for delete tasks.

Ui class: Add delete Ui

This is to display to user delete messages when a successful or failed delete is initiated.

TaskList class: Add delete method

Supporting methods using Java Collections for deleting task in TaskList.