refactor: Update ZBot's addTask method to use ArrayList for tasks
The ZBot class previously used an array to store tasks, but it has been refactored to use an ArrayList instead. This change improves the task display format and allows for dynamic resizing of the task list. The addTask method has been updated to use the add() method of ArrayList to add tasks, and the taskIndex variable has been removed.