Change Duke.java to Nether.java
Duke.java visually prints out "Duke"

The assignment given is to rename the chatbot, and create a greet and
exit message.

Let's rename the file to Nether.java, its new name and create the
greet and exit messages, along with a new visual print out of its name.

"println" statements are the only lines edited as that is the only
requirement for this increment.