Add Echo feature until user inputs "bye"
Chatbot produces visual art of its name, greetings, and goodbye message
when run.

The assignment given is to have chatbot produce echo whenever the
user inputs anything other than "bye".

Let's create a Scanner object to receive user input and create an
infinite loop of echoing the user input and listening for more input. We
used a check inside the infinite loop for a "bye" input to break out of
the loop and produce the goodbye message.