Add "Add" and "List" features and refactored code
Chatbot echoes user input unless the user inputs "bye"

The assignment given is to have the chatbot add whatever the user inputs
into a list of tasks and to add another check for the input "list", to
which the chatbot should return a numbered list of all the tasks the
user has input so far.

Let's create an array of size 100 to store the tasks input by user along
with a counter to keep track of the current index in the array. When
the user inputs "list", a for loop is used to iterate through the array
and print out the entries.

Additionally, an abstraction for printing the horizontal line output is
made and the constants MAX_TASKS, EXIT_COMMAND, and LIST_COMMAND were
created to improve reusability and readability.