Add Mark/Unmark as Done and refactor code
Chatbot add whatever the user inputs into a list of tasks and prints a
numbered list of all the tasks the user has input so far whenever the
user inputs "list".

The assignment given is to have the chatbot have the ability to
mark/unmark tasks as done/not done when the user inputs "mark +
(integer)" or "unmark + (integer)".

Let's create a new class called Task to represent tasks. Task has a
description, index, and an isDone status. Task class also has the
necessary getters and setters. Next, let's implement an extra check for
the user input (i.e. if the user input starts with "mark" or "unmark")
that modifies the task object to be done or not done respectively.

Additionally, an abstraction for displaying the task list is
made and the constants MARK_AS_DONE and MARK_AS_NOT_DONE were
created to improve reusability and readability.

A new Task class was created to encapsulate the data and behavior of
task objects, improve reusability, maintainability, and readability.