Refactor code
The parser does not throw any exceptions when invalid arguments are given.
It relies on what the parser returns when parsing in a user input. If there
were invalid arguments, it will return a Warning executable. This is
not the best way to do it. It should throw exceptions when the input is not
correct, and it should be catched by the bot.

Let's implement it as such. The parser will now throw an exception due to user
input. YihuiBot will catch these exceptions and act accordingly.