Add ToDos, Events, Deadlines
Chatbot adds whatever the user inputs into a list of tasks and prints a
numbered list of all the tasks the user has input so far whenever the
user inputs "list" along with its done/not done status.

The assignment given is to have the chatbot have the ability to
recognize what type of task is added, whether it is ToDos, Deadline, or
Event type tasks and have the task type displayed when listing the
tasks.

Let's create new subclasses for the Task class to represent the 3 types
of tasks. DeadlineTask has a by attribute and EventTask has from and to
attributes. Each of these Task subclasses have their own constructor and
toString methods. Next, let's implement a method that parses the user
input into the correct number of parts to instantiate the respective
tasks with.

Inheritance was used to support multiple task types since the different
tasks have some similarity between them.