Add Error handling and text UI testing
Chatbot adds one of the 3 types of tasks the user inputs into a
numbered list. The tasks can be marked as done or not done, and user can
input "list" to view all the tasks written so far.

The assignment given is to have the chatbot handle errors, especially
when the user enters an unexpected input that the chatbot is not
programmed to handle.

Let's create a new exception called NetherException that can produce the
corresponding error message. Additionally, an automated testing has been
implemented.