Add delete task
Chatbot adds one of the 3 types of tasks the user inputs into a
numbered list. The tasks can be marked as done or not done, and user can
input "list" to view all the tasks written so far. Chatbot can also
handle various inputs from the user.

The assignment given is to have the chatbot able to receive delete
commands to delete a task from the list.

Let's add an extra check to the chatbot for a "delete" command to then
remove the task from the list