Add storage-specific String method for task classes
Tasks have a "toString" method that returns a String representation
of the task's information.

However, tasks need a separate way to format their information as a
String for storage purposes, which is different from the format
used in the UI.

Let's,
* create a new field in Task named "typeIcon" to generalize task
types
* implement a new method that returns a String of the task’s
information formatted for storage

This separation ensures that the data stored in files is consistent
with the expected format during loading, while keeping the UI
representation clean and focused on display purposes.