Implement command parsing and basic commands
This commit introduces a new Parser class and implements basic commands.

Changes include:
- Add Parser class to read user input and handle Add,"bye" and List
- Implement ExitCommand to end chatbot on "bye" command
- Implement ListCommand to display tasks

These changes improves the structure by separating the logic for
handling user inputs and command classes for various actions. This will
allow for new actions to be easily added in the future.

The AddCommand, ExitCommand and ListCommand provides the basic chatbot
functionality
of adding tasks, exiting the program and displaying the tasklist.