Update FishmanException to handle more exceptions
This commit introduces EmptyListException, MissingArgumentException and
NumberFormatException.

EmptyListException manage cases where a task list is empty and
a mark/unmark command is issued. This ensures that the user receives an
error message when they attempt to mark/unmark tasks in an empty list.

MissingArgumentException checks and ensures that each command and the
respective information required is keyed in by the user. This ensures
that the user receives an error message when there is missing
information.

NumberFormatException manages cases where integer values are expected
but are not inputted by the user. This ensures that the user receives an
error message and can make amends to the command.