MoiMoi.java: add storage functionality
MoiMoi is unable to persist tasks between sessions.

Persistent storage is an expected task management feature, as users need
to save tasks and load them in future sessions.

Let's,
* create a class named "Storage" to encapsulate the data file used
to store tasks, along with methods for loading and saving. If the data
file does not exist, a new one will automatically be created
* modify MoiMoi.java to utilize the Storage class for task persistence

By centralizing storage logic in a class, the codebase becomes cleaner
and easier to maintain. Users are enabled to save and load tasks between
program runs.