Update Parser,TaskList and Ui to handle deletion
This commit updates Parser and TaskList class to contain logic for
deletion of task from the tasklist given the correct command.
This ensures that fishman bot is able to parse and execute the delete
command correctly.

The Ui class is updated to handle the display of task deletion, allowing
users to receive the appropriate feedback.