Add support for updating the list
The list that Vecrosen provides is static.
The speak() function doesn't allow text to be added before custom
inputs without using string concatenation.

If the user's input is large, string concatenation would have a large
time complexity, even if the string we wish to add is short.

Let's add anything the user says to the list and update speak() to
accept a prefix which is printed before the text.