Rename main class to 'HypeBot'
Main class name is the default 'Duke'.

The lack of a distinct identity between this new chatbot
and the original Duke makes it hard to differentiate
between the two.

Giving this chatbot a distinct identity allows us to
have an emotional attachment and dedication to
producing a higher quality product. It also allows
users to form an emotional attachment to the
chatbot so they can enjoy spending time interacting
with it.

Let's rename this chatbot 'HypeBot'.

Renaming this chatbot to reflect its desired personality
of a cheerful, motivating companion sets users'
expectations on how their interactions with the chatbot
are going to feel.