Add mark and unmark function and implement classes
The main program is handling all functionalities. This includes
related behaviors.

The related behaviors can be encapsulated in a class which also reduces
code clutter

We should group the related functions of the task list and the task into
it's own classes.

This allows for better modularity of the code for future purposes and
also reduces code clutter.