Add greeting and exit methods
There is only a greeting that is triggered in the
main() method.

The lack of an exit and the lack of separation
between a greeting and an exit complicate the
start and end of the conversation between the
user and the bot.

Extracting the greeting and exit behaviour
each allows us to control the interface at the
start and end of the conversation between the
user and the bot.

Let's make a greet() method that welcomes the
user and and exit() method that wishes them
goodbye.

Separating the greeting and exit message ensures
they can be called separately.