Add support for echo
HypeBot has no support for accepting user
input or reacting to it.

HypeBot needs to be able to accept user
input and have some kind of input handling
so that when the user types in some form of
exit command, HypeBot is able to act accordingly
by calling the exit() method.

Let's use a Scanner and a while loop that
constantly checks whether the exit
command 'bye' in any form of
capitalisation to exit, otherwise
repeat the user's input allows us to
achieve basic echo functionality.

Using a temporary boolean variable as
the condition for the input loop
allows us to store the command entered,
process it into lowercase so we don't
have to test for all combinations which
the command 'bye' can be entered.