Allowed Long-term Storage
The list created inside of the CLI is only able to keep track of one
session and the list is destroyed after the session ends.

This makes the app useless unless the user keeps it open all the
time and if the app is deleted, the user loses all access to what
he/she has done.

Lets create a Storage class to enable long term storage and give the
user the power to decide when to delete tasks even after closing the
session.

Long term storage makes the app more robust in case of accidental
closing of the CLI app and eventually the real app.