Add load and modify delete functionality
The implementation of the long-term storage is incomplete due to
inability to load and delete properly from the long term storage
file.

This makes the app unable to access the information it has stored
and essentially be the same as before the long term storage.

Lets add load and delete from file functionality. Lets also add a
clear functionality to aid with the delete functionality.

The long term storage will now be completely accessible by the app
and be able to support memory between sessions.