Add Check if Storage Directory exists
The program only checked if the file existed not the directory.

If the directory did not exist, the program would have thrown an
error on the first opening of the app itself.

Lets add a check to see if the directory exists when loading the app
at the start of every session.

This makes the system more robust on initial opening and even if the
directory was deleted by accident.