Move general classes to moimoi package
General classes are not organized under any package.

This lack of structure makes the codebase hard to navigate and maintain
as it grows.

Let's,
* move general classes into the "moimoi" package
* edit runtest.bat to include .java files from all packages, ensuring
that they are compiled and available during testing

This ensures that all classes related to MoiMoi are under the "moimoi"
package, with general classes in the root and grouped classes in sub
packages, improving clarity and maintainability.