Separate into packages
The files are not divided into proper packages.

This makes it difficult for any reader to understand which types are
related, where to find which type of classes. It also can cause
namespace conflicts if the number of classes were to increase and
does not encapsulate data well within certain types of files.

Lets separate the files into packages, with the main package being
mortal_reminder, and inside that package we will have 4 subpackages,
namely backend, commands, io and tasks. The main method will remain
inside of the main package as a standalone file.

With these changes, the backend, the tasks, command types and the
user interaction files will all be separated within their own
package folders and the programmer can easily find and change files
they need to in an intuitive manner.