Add gradle support
The app needs to be built and tested manually right now.

Testing and building are two of the things which can be automated
in a project to lighten the load of a programmer as well as ensure
build consistency.

Lets add gradle support for the project.

With gradle support, right now, we can build the project automatically
and run it automatically.