Add JUnit Test for TaskList getTask method
We have been testing the entire application in entirety so far, and
not specific methods to ensure it is not buggy.

This makes it difficult to pinpoint sources of error in the code and
increases the overall debugging time.

Lets create JUnit tests to test the getTask method inside of
TaskList.java. This method is integral to the application because
it is a key part of data storing and manipulation. If this method is
buggy, it may be easy to notice in the results but time-consuming to
check.

Thus, lets create an automated method to test this method to save
debugging time in future.