Add Storage.java test for loadTaskListFromFile
We have been testing the entire application in entirety so far, and
not specific methods to ensure it is not buggy.

This makes it difficult to pinpoint sources of error in the code and
increases the overall debugging time.

Lets create JUnit tests to test the loadTaskListFromFile method
inside of Storage.java.

This method is integral to the application because it is ensures the
long-term memory of the app. If this method is buggy, it may not be
very obvious at first glance, unless tested further in depth.

Thus, an automated method to test this method was created to save
debugging time in future.