Refactor code
The commands in YihuiBot are static.

Users should be able to instantiate the bot whenever they please.

Let's make these methods static.
Refactor YihuiBot to mainly two functions, the constructor and run().
The constructor will instantiate a Ui, Storage, and TaskList class.

Let's create the Ui class.
This class is responsible for all user interface, parsing the user's input
to generate an Executable, as well as formatting the message of YihuiBot
nicely. Contains methods such as prettyPrint and warningPrint. It instantiates
the Parser for generating Executables.

Let's create the Storage class.
This class is responsible for reading the data in specified file and
converting them into a TaskList. It is also responsible for writing the tasks
from TaskList into file. Employs a TaskReader and TaskWriter for each
scenarios.

Let's create the TaskList class. This class extends the ArrayList<Task>
and is used to store the list of tasks created by user. This TaskList would
then be passed to Storage as argument when writing the tasks into file.

Create many new Exceptions, one package for taskformatting exceptions,
one package for parsing exceptions, and one package executable exceptions.