Add Javadocs
All methods were not explained adequately inside the code.

This can lead to future programmers being unable to understand how
to modify the existing code making the code difficult to be improved
on.

Lets add javadocs to as many methods and classed as possible
explaining their functionality succinctly.

This will allow other programmers reading the code to understand and
work with the code or even create tests for the code more easily.