Add 'Save' features to existing project
Project supports simple CRUD functionalities, but the changes made to the listed tasks are not permanently saved onto the hard drive of the user's device.

This prevents previous changes made to the task list from being 'remembered' by the program, rendering the program useless should the user wish to save past tasks.

This commit introduces two new methods:
* getFileContents
* updateFileContents
getFileContents retrieves the information of all previously made tasks (todos, deadlines, and events) from a CSV file (which serves as the database) and returns an ArrayList<Task> containing all these tasks, allowing for the program to interact with past tasks.
updateFileContents iterates through the ArrayList<Task> and writes each and every task onto the CSV file upon being called. It is placed at the end of every CRUD method found in the program.

Now the program is able to output previously created tasks through reading the contents of the 'database.csv' file!