Formatted code to follow a consistent code style
There is some inconsistency in the import lines and the code style
for some of the files.

This makes the code slightly unreadable to the user.

Lets reformat the code using the IDE automation and follow the
recommended coding style in the CS2103 coding style recommendations.

This will make the code alot more readable to the user.