Add find functionality
The list of tasks an actual user will work with is usually rather
large.

This makes it sometimes difficult to find specific task that you have
added in before.

Lets add search functionality for users to be able to use to find
all tasks with similar description to what they have searched for.

This will make the app more easy for users to use in a real
application.