Modified mark and unmark
Marking and Unmarking tasks are not captured in the storage.

If a user were to mark the task and close the app, the change would
not be captured inside of the storage file and would not be reflected
in the app the next time it is turned on.

Lets add the refreshStorage method to the mark and unmark methods to
ensure the changes are kept each time the app is closed.

That should fix any long term memory loss issues with the mark and
unmark methods.