Add TaskList class
Chatbot writes manages task list and updating file storage directly.

This results in poor abstraction and encapsulation,
as it makes the main driver code responsible for
ensuring that task list and file storage are synchronised.

Creating a TaskList class allows us to update the storage
each time the task list and its elements are modified.

Let's make the TaskList class responsible for updating
the storage.

This enforces better abstraction, as the driver code
does not need to be concerned with managing storage.