Add UI class
Chatbot manages formatting console messages directly.

This breaks abstraction and encapsulation,
as it makes the main driver code responsible for directly
managing the UI.

Extracting storage read/writes allow us to
isolate all UI operations.
We can also use polymorphism to replace the CommandLineUI class
with another UI in the future.

Let's put all UI operations in CommandLineUI class,
which manages formatting the messages to display in the console.

This enforces better abstraction, as the driver code
does not need to be concerned with managing display.