Add Save data
Chatbot is initialized with a blank task list every time.

The assignment given is to have the chatbot able to save task list data
and load task list from existing data (if any) upon initialization.

Let's create a new Storage class that handles the saving and loading of
tasks from the data file. Also let's implement a method toSaveFormat
for each task type that processes the existing task list data into a
form that is easily processed later on for loading.