Feature: Save current tasklist to persistent storage on local machine after every change
Previously, the tasklist only existed on local memory and would be erased each time the chatbot was restarted.
Currently, there is already a `./data/save.txt` file storing a representation of a tasklist, and the chatbot will
load up the saved tasks in this list upon startup. But this list can only be manually editted.

It would be better if the save file is automatically updated programatically after every change the user made to
the in memory task list while the chatbot was running.

Let's add an additional function `toSaveRepresentation` to the `Task` class and it's subclasses (`Todo`, `Deadline` and `Event`)
which will generate a String save representation to be stored in the save file. At the start of the user input loop
in the chatbot, Let's store the save representation of every task in the taskList into the save file.