Include Date Functionality
Previously, all tasks that had dates (i.e: Deadline and Event) were simply using strings to format dates.

Since this would not allow for more functionality for later features, and is also prone to errors, this is suboptimal and should be fixed to allow for date time functionality.

Let's update our product to allow for this functionality, and to hit Level-8