Add feature to write inputs to a data file and read from the data file when program starts.
Previously, there is no feature to save inputs from users.

Having this feature makes it more user friendly as users do not have to manually input their tasks every time the program runs.

Save the data from the user input as [Task type] | [IsDone] | [Task Name] to make it easier to query the data. For Event and Deadline data, they have additional columns to store the deadline date and the event from and to data respectively.