Implement Level 7
Gave chatbot the ability to save the tasks in the hard disk automatically whenever the task list changes.

Gave chatbot the ability to load the data from the hard disk when the chatbot starts up.

Added toFileFormat method for Tasks for easier file storage.

Added test cases for to test for loading and storing of data.