feat: add data saving system
The program lacks the ability to save task, resulting in data loss when the exiting program.

Data need to be persisted and able to restore when user restarts the program.

Let's add a saving system to store user data in a text file.

Text file is used to store the data for easy access.