Feature: Add datetime parsing and formatting for `Deadline` and `Event` classes
Previously, the DateTime for the Deadline and Event classes were stored as string. This made it
not very clear on the DateTime and also means we cannot implement features such as search by DateTime
in the future.

Let's standardise the DateTime input to YYYY-MM-DD HH:mm (24 hour format). We can then parse this into
a Java `LocalDateTime` object and format it into a more clear and informative string.